backend/requirements.txt
fastapi
celery
redis
requests
python-dotenv
openai
pandas
beautifulsoup4
playwright
Pillow
uvicorn
python-multipart
woocommerce